requirements.txt
open_clip_torch>=2.24.0
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
chromadb>=0.4.22
matplotlib>=3.8.0
ipywidgets>=8.1.0
jupyter>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.66.0
flask>=3.0.0
